requirements.txt
python-dotenv==1.1.1
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-huggingface==0.3.0
beautifulsoup4==4.13.4
sentence_transformers==5.0.0
streamlit==1.46.1